requirements.txt
pyTelegramBotAPI==4.7.0
python-telegram-bot==13.13